reverted unrelated code cleanup change
